refactor cached texture class
