+++ client/unity/simple-chat/Assets/Script/SimpleChat/Domain/Service/CachedTexture.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f349aba6fad584f4eb7f36e4ecb0462a
+guid: 75938201e21de403c8e4a82c5b19301f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
